NO_JIRA - refactor CiagInductionEventBuilder to InboundEventBuilder in correct package
