+++ src/testFixtures/kotlin/uk/gov/justice/digital/hmpps/educationandworkplanapi/app/messaging/InboundEventBuilder.kt
@@ -1,11 +1,6 @@
-package uk.gov.justice.digital.hmpps.educationandworkplanapi.service.events
+package uk.gov.justice.digital.hmpps.educationandworkplanapi.app.messaging
 
 import uk.gov.justice.digital.hmpps.educationandworkplanapi.aValidPrisonNumber
-import uk.gov.justice.digital.hmpps.educationandworkplanapi.app.messaging.AdditionalInformation
-import uk.gov.justice.digital.hmpps.educationandworkplanapi.app.messaging.EventType
-import uk.gov.justice.digital.hmpps.educationandworkplanapi.app.messaging.Identifier
-import uk.gov.justice.digital.hmpps.educationandworkplanapi.app.messaging.InboundEvent
-import uk.gov.justice.digital.hmpps.educationandworkplanapi.app.messaging.PersonReference
 import java.time.Instant
 import java.util.UUID
